user can now select a profile picture

+++ stackoverflow/src/main/java/com/stackoverflow/service/recoverypassword/CodeVerificationServiceImpl.java
@@ -8,6 +8,9 @@ import com.stackoverflow.repository.CodeVerificationRepository;
 import com.stackoverflow.repository.UserRepository;
 import com.stackoverflow.service.MailService;
 import jakarta.transaction.Transactional;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.Validator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,10 +20,11 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.security.SecureRandom;
 import java.time.LocalDateTime;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
-public class CodeVerificationServiceImpl implements CodeVerificationService{
+public class CodeVerificationServiceImpl implements CodeVerificationService {
     @Autowired
     private CodeVerificationRepository codeVerificationRepository;
 
@@ -30,13 +34,15 @@ public class CodeVerificationServiceImpl implements CodeVerificationService{
     @Autowired
     private MailService mailService;
 
+    private final Validator validator;
+
     private final PasswordEncoder passwordEncoder;
 
-    public String generateCode(){
+    public String generateCode() {
         String numbers = "0123456789";
         SecureRandom random = new SecureRandom();
         StringBuilder code = new StringBuilder();
-        for(int i=0; i<4; i++){
+        for (int i = 0; i < 4; i++) {
             code.append(numbers.charAt(random.nextInt(numbers.length())));
         }
         return code.toString();
@@ -69,7 +75,8 @@ public class CodeVerificationServiceImpl implements CodeVerificationService{
     public void verifyCodeAndChangePassword(PasswordResetDto passwordResetDto) {
         CodeVerification codeVerification = codeVerificationRepository
                 .findByEmailAndCode(passwordResetDto.getEmail(), passwordResetDto.getCode())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid code or email not found"));
+                .orElseThrow(
+                        () -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid code or email not found"));
 
         if (codeVerification.getDateExpiration().isBefore(LocalDateTime.now())) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Code expired");
@@ -78,7 +85,17 @@ public class CodeVerificationServiceImpl implements CodeVerificationService{
         User user = userRepository.findByEmail(passwordResetDto.getEmail())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
 
+        String regex = "^(?=.*\\d)(?=.*[\\u0021-\\u002b\\u003c-\\u0040])(?=.*[A-Z])(?=.*[a-z])\\S{8,16}$";
+        if (!passwordResetDto.getNewPassword().matches(regex)) {
+            throw new IllegalArgumentException("New password does not meet security requirements");
+        }
+
         user.setPassword(passwordEncoder.encode(passwordResetDto.getNewPassword()));
+
+        Set<ConstraintViolation<User>> violations = validator.validate(user);
+        if (!violations.isEmpty())
+            throw new ConstraintViolationException(violations);
+
         userRepository.save(user);
 
         codeVerificationRepository.deleteByEmail(passwordResetDto.getEmail());
